feat(contract_manager): relay a price update before proposing the SVM migration
The propose script now pulls an update from the Hermes that is live today and
relays it through each chain's receiver before it proposes anything, so the
execute script's post-migration relay has a baseline to be read against.

The relay moves out of the execute script into the shared module, and returns a
description rather than logging, so both halves can use it. A failure is
reported rather than fatal — a chain that cannot relay a price today is still
worth migrating.

The upgrade-authority reads in the state dump no longer abort the run either: a
program that is not deployed on the chain is reported in place of its key.
